resourceloader: Simplify isCompatible() parameter in startup.js
This is a private function (now marked as such) only ever called
once in prod code (a few lines further down). It is exposed
to testing as well, which is why it has a parameter.

Call it with a parameter within startup.js as well, so that we can
remove the default handling and use the function the same way in
prod as the way we use it in tests.

Change-Id: Ie9db60e417cca29f86cea20359e880cb1fa20197
(cherry picked from commit 5157ccb16af58c4a58ce01a22ba6d2e33935c52b)