parser: use "-" for revision ID for non-preview edit filter parse during save
This avoids a double parse when the edit stash is not used,
which can be confirmed via the SaveParse log for a page
using {{REVISIONID}} when edit stashing is disabled. This
now matches the reuse for the edit stash hit case.

Change-Id: I405c39d4d7ac04e39fbdfe400f73238b734c7833
(cherry picked from commit d256b472f73956ee8e2503e0254a1107baa1f00a)